extract UpdateConverationMemberLeave
